requirements.txt
pandas~=2.0.3
streamlit==1.29.0
watchdog==3.0.0
numpy~=1.26.2
scikit-learn~=1.3.2
Pillow~=9.5.0
